wmfgerrit: tweak applied styles
Set script to run in the document-idle phase to see if this makes styles
actually show in the browser. May have to get more fancy with this as
the UI seems to be at least partially drawn using JS which appears to
mess with the ability of Greasemonkey to style it.

Also try more forcefully to suppress the text site id.